added: RequestIntervalMiddleware, ProxyMiddleware, GuzzleClientFactory and guzzle parameters constant classes refactoring replaced async requests with Pool for downloading images (in progress) added rejected requests and retries

+++ src/Dto/Manga/MangaChaptersListItemDto.php
@@ -5,7 +5,7 @@ namespace App\Dto\Manga;
 class MangaChaptersListItemDto
 {
     public function __construct(
-        public int $number,
+        public float $number,
         public int $volume
     )
     {
